pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "forge"
version = "0.1.0"
description = "FORGE Test Harness - CI-first, reproducible and observable performance & scale testing"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Red Hat PSAP Team", email = "psap@redhat.com"}
]
keywords = ["testing", "performance", "scale", "openshift", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "plotly>=5.17.0",
    "dash>=2.14.0",
    "dash-bootstrap-components>=1.5.0",
    "pyyaml>=6.0",
    "jsonschema>=4.19.0",
    "structlog>=23.1.0",
    "click>=8.1.0",
    "requests",
    "slack_sdk==3.35.0",
    "ansible", "fire",
    "pyjwt",
    "jsonpath_ng",
    "pydantic>=2.0",
    "yq",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "black>=23.7.0",
    "ruff>=0.0.287",
    "pre-commit>=4.0",
]
docs = [
    "sphinx>=7.1.0",
    "sphinx-rtd-theme>=1.3.0",
]
caliper = [
    "opensearch-py>=2.4.0",
    "boto3>=1.28.0",
    "mlflow>=2.10.0",
]
matrix_benchmarking = [
    "pandas",
    "dash",
    "plotly",
    "kaleido==0.2.1",
    "fire",
    "pyyaml",
    "BeautifulSoup4==4.11.*",
    "prometheus_api_client",
    "pydantic",
    "opensearch-py",
    "boto3",
    "pandas",
]

[project.scripts]
caliper = "projects.caliper.cli.main:run_cli"

[tool.setuptools.packages.find]
where = ["projects"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["projects/core/tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
